-Cementation
Cementation is the process of lowering the expandability of the API to
improve efficiency and likelihood to not run into errors.

This is primarily being achieved by removing the dynamic loading system
I have for the emotions and traits. 

I do still plan to create a "lite" version for those who do still wish
to create their own emotions and traits system.